fix: 위치 타입수정, 게임시간 수정

+++ backend/game_service/src/main/java/com/s10p31a709/game/api/room/entity/Player.java
@@ -19,9 +19,9 @@ public class Player {
     @JsonProperty("isSeeker")
     private Boolean isSeeker;
 
-    private Integer[] position;
+    private Double[] position;
 
-    private Integer[] direction;
+    private Double[] direction;
 
     @JsonIgnore
     private String sessionId;
